changed my commit to adhere to suggestions made
