+++ frontend/src/components/shared/ui/tabs.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import * as TabsPrimitive from "@radix-ui/react-tabs";
 import * as React from "react";
